requirements.txt
Flask
Flask-SQLAlchemy
Flask-security-too
Flask-caching
Celery[redis]
bleach
redis
redis-server
requests
smtplib